pyproject.toml
[project]
name = "django-carros-api"
version = "0.1.0"
description = ""
authors = [
    {name = "lphillipe",email = "luisphillipe200@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "groq (>=1.1.2,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
